requirements.txt
discord.py>=2.0.0
googletrans==4.0.0-rc1
